Rename NoBatchReportEntryWrapper
