+++ backend/de.metas.banking/de.metas.banking.camt53/src/main/java/de/metas/banking/camt53/wrapper/v02/AccountStatement2Wrapper.java
@@ -151,7 +151,7 @@ public class AccountStatement2Wrapper extends AccountStatementWrapper
 	@NonNull
 	private IStatementLineWrapper buildNoBatchReportEntryWrapper(@NonNull final ReportEntry2 reportEntry)
 	{
-		return NoBatchReportEntry2Wrapper.builder()
+		return BatchReportEntry2Wrapper.builder()
 				.currencyRepository(getCurrencyRepository())
 				.entry(reportEntry)
 				.build();
