+++ backend/de.metas.banking/de.metas.banking.camt53/src/main/java/de/metas/banking/camt53/wrapper/v02/BatchReportEntry2Wrapper.java
@@ -20,21 +20,20 @@
  * #L%
  */
 
-package de.metas.banking.camt53.wrapper.v04;
+package de.metas.banking.camt53.wrapper.v02;
 
 import com.google.common.collect.ImmutableList;
-import de.metas.banking.camt53.jaxb.camt053_001_04.ActiveOrHistoricCurrencyAndAmount;
-import de.metas.banking.camt53.jaxb.camt053_001_04.AmountAndCurrencyExchange3;
-import de.metas.banking.camt53.jaxb.camt053_001_04.AmountAndCurrencyExchangeDetails3;
-import de.metas.banking.camt53.jaxb.camt053_001_04.CurrencyExchange5;
-import de.metas.banking.camt53.jaxb.camt053_001_04.DateAndDateTimeChoice;
-import de.metas.banking.camt53.jaxb.camt053_001_04.EntryDetails3;
-import de.metas.banking.camt53.jaxb.camt053_001_04.EntryTransaction4;
-import de.metas.banking.camt53.jaxb.camt053_001_04.InterestRecord1;
-import de.metas.banking.camt53.jaxb.camt053_001_04.RemittanceInformation7;
-import de.metas.banking.camt53.jaxb.camt053_001_04.ReportEntry4;
-import de.metas.banking.camt53.jaxb.camt053_001_04.TransactionInterest3;
-import de.metas.banking.camt53.wrapper.NoBatchReportEntryWrapper;
+import de.metas.banking.camt53.jaxb.camt053_001_02.ActiveOrHistoricCurrencyAndAmount;
+import de.metas.banking.camt53.jaxb.camt053_001_02.AmountAndCurrencyExchange3;
+import de.metas.banking.camt53.jaxb.camt053_001_02.AmountAndCurrencyExchangeDetails3;
+import de.metas.banking.camt53.jaxb.camt053_001_02.CurrencyExchange5;
+import de.metas.banking.camt53.jaxb.camt053_001_02.DateAndDateTimeChoice;
+import de.metas.banking.camt53.jaxb.camt053_001_02.EntryDetails1;
+import de.metas.banking.camt53.jaxb.camt053_001_02.EntryTransaction2;
+import de.metas.banking.camt53.jaxb.camt053_001_02.RemittanceInformation5;
+import de.metas.banking.camt53.jaxb.camt053_001_02.ReportEntry2;
+import de.metas.banking.camt53.jaxb.camt053_001_02.TransactionInterest2;
+import de.metas.banking.camt53.wrapper.BatchReportEntryWrapper;
 import de.metas.currency.CurrencyCode;
 import de.metas.currency.CurrencyRepository;
 import de.metas.money.Money;
@@ -53,19 +52,19 @@ import java.util.TimeZone;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static de.metas.banking.camt53.jaxb.camt053_001_04.CreditDebitCode.CRDT;
+import static de.metas.banking.camt53.jaxb.camt053_001_02.CreditDebitCode.CRDT;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
-public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
+public class BatchReportEntry2Wrapper extends BatchReportEntryWrapper
 {
 	@NonNull
-	ReportEntry4 entry;
+	ReportEntry2 entry;
 
 	@Builder
-	private NoBatchReportEntry4Wrapper(
+	private BatchReportEntry2Wrapper(
 			@NonNull final CurrencyRepository currencyRepository,
-			@NonNull final ReportEntry4 entry)
+			@NonNull final ReportEntry2 entry)
 	{
 		super(currencyRepository);
 		Check.assume(Check.isEmpty(entry.getNtryDtls()) || entry.getNtryDtls().size() == 1, "Batched transactions are not supported!");
@@ -74,12 +73,12 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	}
 
 	@NonNull
-	public Optional<EntryTransaction4> getEntryTransaction()
+	public Optional<EntryTransaction2> getEntryTransaction()
 	{
 		return Optional.of(entry.getNtryDtls())
 				.filter(list -> !list.isEmpty())
 				.map(list -> list.get(0))
-				.map(EntryDetails3::getTxDtls)
+				.map(EntryDetails1::getTxDtls)
 				.filter(list -> !list.isEmpty())
 				.map(list -> list.get(0));
 	}
@@ -98,10 +97,9 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	public Optional<Money> getInterestAmount()
 	{
 		final ActiveOrHistoricCurrencyAndAmount activeOrHistoricCurrencyAndAmount = Optional.ofNullable(entry.getIntrst())
-				.map(TransactionInterest3::getRcrd)
 				.filter(list -> !list.isEmpty())
 				.map(list -> list.get(0))
-				.map(InterestRecord1::getAmt)
+				.map(TransactionInterest2::getAmt)
 				.orElse(null);
 
 		if (activeOrHistoricCurrencyAndAmount == null || activeOrHistoricCurrencyAndAmount.getValue() == null)
@@ -119,7 +117,7 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	public Optional<BigDecimal> getCurrencyRate()
 	{
 		return getEntryTransaction()
-				.map(EntryTransaction4::getAmtDtls)
+				.map(EntryTransaction2::getAmtDtls)
 				.map(AmountAndCurrencyExchange3::getCntrValAmt)
 				.map(AmountAndCurrencyExchangeDetails3::getCcyXchg)
 				.map(CurrencyExchange5::getXchgRate);
@@ -146,8 +144,8 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	public String getDbtrNames()
 	{
 		return entry.getNtryDtls().stream()
-				.flatMap(entryDetails3 -> entryDetails3.getTxDtls().stream())
-				.map(EntryTransaction4::getRltdPties)
+				.flatMap(entryDetails1 -> entryDetails1.getTxDtls().stream())
+				.map(EntryTransaction2::getRltdPties)
 				.filter(party -> party != null && party.getDbtr() != null)
 				.map(party -> party.getDbtr().getNm())
 				.filter(Check::isNotBlank)
@@ -159,8 +157,8 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	public String getCdtrNames()
 	{
 		return entry.getNtryDtls().stream()
-				.flatMap(entryDetails3 -> entryDetails3.getTxDtls().stream())
-				.map(EntryTransaction4::getRltdPties)
+				.flatMap(entryDetails1 -> entryDetails1.getTxDtls().stream())
+				.map(EntryTransaction2::getRltdPties)
 				.filter(party -> party != null && party.getCdtr() != null)
 				.map(party -> party.getCdtr().getNm())
 				.filter(Check::isNotBlank)
@@ -179,8 +177,8 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	protected String getUnstructuredRemittanceInfo(@NonNull final String delimiter)
 	{
 		return String.join(delimiter, getEntryTransaction()
-				.map(EntryTransaction4::getRmtInf)
-				.map(RemittanceInformation7::getUstrd)
+				.map(EntryTransaction2::getRmtInf)
+				.map(RemittanceInformation5::getUstrd)
 				.orElseGet(ImmutableList::of));
 	}
 
@@ -189,7 +187,7 @@ public class NoBatchReportEntry4Wrapper extends NoBatchReportEntryWrapper
 	protected String getLineDescription(@NonNull final String delimiter)
 	{
 		final String trxDetails = getEntryTransaction()
-				.map(EntryTransaction4::getAddtlTxInf)
+				.map(EntryTransaction2::getAddtlTxInf)
 				.filter(Check::isNotBlank)
 				.orElse(null);
