+++ backend/de.metas.banking/de.metas.banking.camt53/src/main/java/de/metas/banking/camt53/wrapper/v04/AccountStatement4Wrapper.java
@@ -151,7 +151,7 @@ public class AccountStatement4Wrapper extends AccountStatementWrapper
 	@NonNull
 	private IStatementLineWrapper buildNoBatchReportEntryWrapper(@NonNull final ReportEntry4 reportEntry)
 	{
-		return NoBatchReportEntry4Wrapper.builder()
+		return BatchReportEntry4Wrapper.builder()
 				.currencyRepository(getCurrencyRepository())
 				.entry(reportEntry)
 				.build();
